Break generators into a separate file.
